gametest/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.7.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.5
sqlparse==0.4.3